Migration case & inputv2 flaky fix
